individual_bookmaker_service/requirements.txt
# Django và DRF
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.0.0,<5.0.0
django-filter>=23.0,<24.0

# Database
psycopg2-binary>=2.9.0,<3.0.0
redis>=4.5.0,<5.0.0

# Authentication
djangorestframework-simplejwt>=5.2.0,<6.0.0

# Utilities
python-decouple>=3.8,<4.0.0
celery>=5.3.0,<6.0.0

# Development và Testing
pytest>=7.4.0,<8.0.0
pytest-django>=4.5.0,<5.0.0
coverage>=7.2.0,<8.0.0
black>=23.0,<24.0
flake8>=6.0.0,<7.0.0

# Production
gunicorn>=21.0.0,<22.0.0
whitenoise>=6.5.0,<7.0.0

# Monitoring và Logging
sentry-sdk>=1.28.0,<2.0.0
